Add contification step
This currently breaks tests due to continuation marks being cleared
via what is now an inbalance in continuation creation/applications.
